+++ src/fonctions_graphes.c
@@ -12,9 +12,9 @@ Graph graph_of_SimpleMap(SimpleMap map){
 
 	Graph graph = create_Graph( N * M );
 
-	for(int i = 0; i < N; i++){
+	for(uint16_t i = 0; i < N; i++){
 
-		for(int j = 0; j < M; j++){
+		for(uint16_t j = 0; j < M; j++){
 
 			if(!map[i][j]){
 				if( i > 0 && !map[i-1][j]) add_arrete(&graph, i*N + j, (i-1)*N + j);
@@ -30,8 +30,8 @@ Graph graph_of_SimpleMap(SimpleMap map){
 
 Couple resultat_graphe_simple(SimpleMap map, Coordonnee coordonneeBegin, Coordonnee coordonneeEnd){
 
-	unsigned int begin = coordonneeBegin.x * N + coordonneeBegin.y;
-	unsigned int end = coordonneeEnd.x * N + coordonneeEnd.y;
+	uint16_t begin = coordonneeBegin.x * N + coordonneeBegin.y;
+	uint16_t end = coordonneeEnd.x * N + coordonneeEnd.y;
 
 	Graph graph = graph_of_SimpleMap(map);
 
@@ -44,19 +44,19 @@ Couple resultat_graphe_simple(SimpleMap map, Coordonnee coordonneeBegin, Coordon
 		exit(MALLOC_ERROR);
 	}
 
-	for(int i = 0; i < graph.n; i++)
+	for(uint32_t i = 0; i < graph.n; i++)
 		mark[i] = false;
 
-	int* fathers;
+	uint32_t* fathers;
 
-	fathers = calloc( graph.n, sizeof(int));
+	fathers = calloc( graph.n, sizeof(uint32_t));
 
 	if( !fathers ){
 		printf("Unable to allocate the memory for the fathers");
 		exit(MALLOC_ERROR);
 	}
 
-	for(int i = 0; i < graph.n; i++)
+	for(uint32_t i = 0; i < graph.n; i++)
 		fathers[i] = -1;
 
 	Queue queue = create_Queue();
